Feat #10 - View 연결 및 최종 적용

+++ src/main/java/com/jsplan/drp/domain/sys/usermng/service/UserGrpMngService.java
@@ -4,14 +4,12 @@ import com.jsplan.drp.domain.sys.usermng.dto.UserGrpMngRequest;
 import com.jsplan.drp.domain.sys.usermng.dto.UserGrpMngResponse;
 import com.jsplan.drp.domain.sys.usermng.dto.UserGrpMngSearchDto;
 import com.jsplan.drp.domain.sys.usermng.entity.UserGrpMng;
-import com.jsplan.drp.domain.sys.usermng.entity.UserGrpMngDto.UserGrpMngDetailDto;
-import com.jsplan.drp.domain.sys.usermng.entity.UserGrpMngDto.UserGrpMngListDto;
+import com.jsplan.drp.domain.sys.usermng.dto.UserGrpMngDto.UserGrpMngDetailDto;
+import com.jsplan.drp.domain.sys.usermng.dto.UserGrpMngDto.UserGrpMngListDto;
 import com.jsplan.drp.domain.sys.usermng.repository.UserGrpMngRepository;
-import com.jsplan.drp.global.exception.entity.ErrorStatus;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
@@ -35,7 +33,7 @@ public class UserGrpMngService {
      * @param searchDto (조회 조건)
      * @return Page (페이징 목록)
      */
-    public Page<UserGrpMngListDto> selectGrpMngList(UserGrpMngSearchDto searchDto) {
+    public Page<UserGrpMngListDto> selectUserGrpMngList(UserGrpMngSearchDto searchDto) {
         PageRequest pageRequest = PageRequest.of(searchDto.getPageNo(), searchDto.getPageSize());
         return userGrpMngRepository.searchPageList(searchDto.getGrpNm(), pageRequest);
     }
@@ -46,7 +44,7 @@ public class UserGrpMngService {
      * @param request (그룹 정보)
      * @return UserGrpMngDto (그룹 DTO)
      */
-    public UserGrpMngDetailDto selectGrpMngDetail(UserGrpMngRequest request) {
+    public UserGrpMngDetailDto selectUserGrpMngDetail(UserGrpMngRequest request) {
         return userGrpMngRepository.findByGrpCd(request.getGrpCd());
     }
 
@@ -54,48 +52,69 @@ public class UserGrpMngService {
      * <p>그룹 등록</p>
      *
      * @param request (그룹 정보)
-     * @return UserGrpMngResponse (그룹 코드)
+     * @return UserGrpMngResponse (응답 정보)
      */
     @Transactional
-    public UserGrpMngResponse insertGrpMngData(UserGrpMngRequest request) {
-        validateDupGrpMngData(request);
-        UserGrpMng userGrpMng = userGrpMngRepository.save(request.toEntity());
-        return new UserGrpMngResponse(userGrpMng.getGrpCd());
+    public UserGrpMngResponse insertUserGrpMngData(UserGrpMngRequest request) {
+        if (validateDupGrpMngData(request)) {
+            return new UserGrpMngResponse(null, "D");
+        } else {
+            UserGrpMng userGrpMng = userGrpMngRepository.save(request.toEntity());
+            return new UserGrpMngResponse(userGrpMng.getGrpCd(), "S");
+        }
     }
 
     /**
      * <p>중복 그룹 체크</p>
      *
      * @param request (그룹 정보)
+     * @return boolean (중복 여부)
      */
-    private void validateDupGrpMngData(UserGrpMngRequest request) {
-        Optional<UserGrpMng> findData = userGrpMngRepository.findById(request.getGrpCd());
-        findData.ifPresent(findUser -> {
-            throw new DataIntegrityViolationException(ErrorStatus.DUPLICATED_KEY.getMessage());
-        });
+    private boolean validateDupGrpMngData(UserGrpMngRequest request) {
+        Optional<UserGrpMng> optionalUserGrpMng = userGrpMngRepository.findById(request.getGrpCd());
+        return optionalUserGrpMng.isPresent();
+//        optionalUserGrpMng.ifPresent(userGrpMng -> {
+//            throw new DataIntegrityViolationException(ErrorStatus.DUPLICATED_KEY.getMessage());
+//        });
     }
 
     /**
      * <p>그룹 수정</p>
      *
      * @param request (그룹 정보)
-     * @return UserGrpMngResponse (그룹 코드)
+     * @return UserGrpMngResponse (응답 정보)
      */
     @Transactional
-    public UserGrpMngResponse updateGrpMngData(UserGrpMngRequest request) {
+    public UserGrpMngResponse updateUserGrpMngData(UserGrpMngRequest request) {
         UserGrpMng userGrpMng = userGrpMngRepository.findById(request.getGrpCd())
             .orElseThrow(NoSuchElementException::new);
         userGrpMng.update(request);
-        return new UserGrpMngResponse(userGrpMng.getGrpCd());
+        return new UserGrpMngResponse(userGrpMng.getGrpCd(), "S");
     }
 
     /**
      * <p>그룹 삭제</p>
      *
      * @param request (그룹 정보)
+     * @return UserGrpMngResponse (응답 정보)
      */
     @Transactional
-    public void deleteGrpMngData(UserGrpMngRequest request) {
-        userGrpMngRepository.deleteById(request.getGrpCd());
+    public UserGrpMngResponse deleteUserGrpMngData(UserGrpMngRequest request) {
+        if (existsUserMngData(request)) { // 그룹 내 사용자가 존재할 경우
+            return new UserGrpMngResponse(request.getGrpCd(), "F");
+        } else {
+            userGrpMngRepository.deleteById(request.getGrpCd());
+            return new UserGrpMngResponse(request.getGrpCd(), "S");
+        }
+    }
+
+    /**
+     * <p>사용자 확인</p>
+     *
+     * @param request (그룹 정보)
+     * @return boolean (사용자 존재 여부)
+     */
+    private boolean existsUserMngData(UserGrpMngRequest request) {
+        return userGrpMngRepository.existsUserMngByGrpCd(request.getGrpCd());
     }
 }
